feat:데이터 옮김 및 리덕스로 데이터 보냄(#20)
